moved `getEllipsizedText` to text.ts and added unit tests for the same
